+++ src/main/kotlin/com/github/copyinaction/config/OpenApiConfig.kt
@@ -3,7 +3,6 @@ package com.github.copyinaction.config
 import io.swagger.v3.oas.models.Components
 import io.swagger.v3.oas.models.OpenAPI
 import io.swagger.v3.oas.models.info.Info
-import io.swagger.v3.oas.models.security.SecurityRequirement
 import io.swagger.v3.oas.models.security.SecurityScheme
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
@@ -22,7 +21,6 @@ class OpenApiConfig {
                     .description("Smarter Store 백엔드 API 문서")
                     .version("1.0.0")
             )
-            .addSecurityItem(SecurityRequirement().addList(securitySchemeName))
             .components(
                 Components()
                     .addSecuritySchemes(