+++ src/main/kotlin/com/github/copyinaction/controller/AuthController.kt
@@ -13,6 +13,7 @@ import io.swagger.v3.oas.annotations.media.Content
 import io.swagger.v3.oas.annotations.media.Schema
 import io.swagger.v3.oas.annotations.responses.ApiResponse
 import io.swagger.v3.oas.annotations.responses.ApiResponses
+import io.swagger.v3.oas.annotations.security.SecurityRequirements
 import io.swagger.v3.oas.annotations.tags.Tag
 import jakarta.validation.Valid
 import org.springframework.http.ResponseEntity
@@ -21,7 +22,8 @@ import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 
-@Tag(name = "인증 API", description = "사용자 회원가입 및 로그인 API")
+@Tag(name = "인증 API", description = "사용자 회원가입 및 로그인 API (인증 불필요)")
+@SecurityRequirements
 @RestController
 @RequestMapping("/api/auth")
 class AuthController(